♻️ refactor(models): consolidate model name resolution logic
- remove redundant resolve_model_name function from utils
- integrate model name resolution into ModelRegistry class
- update endpoints to use ModelRegistry for resolving model names

🔥 chore(dev_scripts): remove unused test_resolve_name script

- delete test_resolve_name.py script as it's no longer needed
