--- dev_scripts/test_resolve_name.py
@@ -1,10 +1,0 @@
-import os
-
-from argoproxy.models import CHAT_MODELS
-from argoproxy.utils.models import resolve_model_name
-
-MODEL = os.getenv("MODEL", "argo:gpt-4o")
-DEFAULT_MODEL = "argo:gpt-4o"
-
-resolved_name = resolve_model_name(MODEL, DEFAULT_MODEL, avail_models=CHAT_MODELS)
-print(resolved_name)
